AG-Refactor-Tests-stage-1

+++ CastleSiegePom/CastleSiegeApp/src/test/java/com/pilaf/cs/tests/SpringIntegrationTest.java
@@ -6,13 +6,12 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.context.ContextConfiguration;
 
-//@WebAppConfiguration
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @ContextConfiguration
 public class SpringIntegrationTest {
 
 	@LocalServerPort
-	public long port;
+	public int port;
 	
 	@Value("${jwt.header}")
 	public String path;
